Moved Search to Component
