--- src/Pages/SearchPage/SearchPage.css
@@ -1,35 +1,0 @@
-.search-page-search {
-    border-radius: 10px;
-    border-style: solid;
-    border-color: #323232;
-
-    padding: 0.3em;
-    padding-left: 0.8em;
-    margin-bottom: 1em;
-    font-size: 24px;
-    font-family: 'IBM Plex Mono', monospace;
-    width: 97.5%;
-}
-
-.search-page-container {
-    background-color: white;
-    padding: 1em;
-    margin-top: 2em;
-    border-radius: 10px;
-}
-
-.search-results {
-    background-color: #DEDEDE;
-    padding: 10px;
-    display: grid;
-    column-gap: 5px;
-    row-gap: 20px;
-    grid-auto-columns: 200px;
-    grid-template-columns: repeat(auto-fill, minmax(var(--video-preview-width), 1fr))
-}
-
-@media screen and (max-width: 1300px) {
-    .search-page-search {
-        width: 90%;
-    }
-}
